Introduce setTag method for a Span interface and get rid of setTags
Having setTags instead of setTag means that you would have to be aware
of tags that are set in order to add aditional tag to a span. But that is
not the case if we expose an interface to set just a single tag.

All other opentracing libraries also have setTag exposed instead of
setTags so this helps to have more consistency between libraries.

Closes #51